backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.1
ffmpeg-python==0.2.0
pillow==10.2.0
torch --index-url https://download.pytorch.org/whl/cpu
transformers==4.37.2
pydantic==2.5.3
pydantic-settings==2.1.0
sqlite-utils==3.36
